+++ packages/codegen/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Fixed
+
+- Storage entry values decode to `null` if the entry is void.
+
 ## 0.15.0 - 2025-05-15
 
 ### Added